invoice history class id to long

+++ nanopay/src/net/nanopay/invoice/ui/history/InvoiceHistoryView.js
@@ -30,7 +30,7 @@ foam.CLASS({
 
   properties: [
     {
-      class: 'Int',
+      class: 'Long',
       name: 'id',
       documentation: `Please use class when defining data type`
     },
